requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
spacy
sentence-transformers
scikit-learn
